+++ docs/.vitepress/components/radio/group.vue
@@ -1,16 +1,7 @@
 <template>
     <space>
         <FRadioGroup v-model="val" @change="handleChange">
-            <FRadio :value="1">1</FRadio>
-            <FRadio :value="2">22</FRadio>
-            <FRadio :value="3">333</FRadio>
-            <FRadio :value="4">4444</FRadio>
-        </FRadioGroup>
-        <FRadioGroup disabled>
-            <FRadio :value="1">1</FRadio>
-            <FRadio :value="2">22</FRadio>
-            <FRadio :value="3">333</FRadio>
-            <FRadio :value="4">4444</FRadio>
+            <FRadio v-for="i in len" :key="i" :value="i">{{ i }}</FRadio>
         </FRadioGroup>
     </space>
 </template>
@@ -19,10 +10,12 @@ import { ref } from 'vue';
 export default {
     setup() {
         const val = ref();
+        const len = 30;
         const handleChange = (value) => {
             console.log('change:', value);
         };
         return {
+            len,
             val,
             handleChange,
         };